Silent commands
Closes #27

+++ src/main/java/com/telluur/SlapBot/commands/abstractions/AbstractCommand.java
@@ -23,6 +23,9 @@ public abstract class AbstractCommand extends Command {
     private static final Logger logger = LoggerFactory.getLogger("CMD");
     protected SlapBot slapBot;
 
+    protected boolean clearUserIssue = true;
+    protected boolean echoUserIssue = true;
+
     AbstractCommand(SlapBot slapBot) {
         this.slapBot = slapBot;
     }
@@ -33,26 +36,31 @@ public abstract class AbstractCommand extends Command {
         Scan and replace mentions in the command when used in guild.
          */
         if (event.getChannel().getType().equals(ChannelType.TEXT)) {
-            event.getMessage().delete().queue();
-            String filteredCmd = Arrays.stream(event.getMessage().getContentRaw().split("\\s+"))
-                    .map(stringPart -> {
-                        if (stringPart.startsWith("<")) {
-                            List<Member> possibleMembers = FinderUtil.findMembers(stringPart, slapBot.getGuild());
-                            List<Role> possibleRoles = FinderUtil.findRoles(stringPart, slapBot.getGuild());
-                            if (possibleMembers.size() == 1) {
-                                Member m = possibleMembers.get(0);
-                                return String.format("`%s`", m.getEffectiveName());
-                            } else if (possibleRoles.size() == 1) {
-                                return String.format("`%s`", possibleRoles.get(0).getName());
+            if (clearUserIssue) {
+                event.getMessage().delete().queue();
+            }
+
+            if (echoUserIssue) {
+                String filteredCmd = Arrays.stream(event.getMessage().getContentRaw().split("\\s+"))
+                        .map(stringPart -> {
+                            if (stringPart.startsWith("<")) {
+                                List<Member> possibleMembers = FinderUtil.findMembers(stringPart, slapBot.getGuild());
+                                List<Role> possibleRoles = FinderUtil.findRoles(stringPart, slapBot.getGuild());
+                                if (possibleMembers.size() == 1) {
+                                    Member m = possibleMembers.get(0);
+                                    return String.format("`%s`", m.getEffectiveName());
+                                } else if (possibleRoles.size() == 1) {
+                                    return String.format("`%s`", possibleRoles.get(0).getName());
+                                } else {
+                                    return stringPart;
+                                }
                             } else {
                                 return stringPart;
                             }
-                        } else {
-                            return stringPart;
-                        }
-                    })
-                    .collect(Collectors.joining(" "));
-            event.getChannel().sendMessage(String.format("`%s` issued command: %s", event.getMember().getEffectiveName(), filteredCmd)).queue();
+                        })
+                        .collect(Collectors.joining(" "));
+                event.getChannel().sendMessage(String.format("`%s` issued command: %s", event.getMember().getEffectiveName(), filteredCmd)).queue();
+            }
         }
     }
